web/web_test.go
/*
 * Copyright (c) 2023 sixwaaaay.
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *     http://www.apache.org/licenses/LICENSE-2.
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package web

import (
	"net/http"
	"net/http/httptest"
	"testing"

	"github.com/golang-jwt/jwt/v5"
	"github.com/stretchr/testify/assert"
)

func TestMiddleware_WithValidToken(t *testing.T) {
	secret := []byte("secret")
	middleware := Middleware(secret)

	req, _ := http.NewRequest("GET", "/", nil)
	req.Header.Add("Authorization", "Bearer "+createToken(secret))
	rr := httptest.NewRecorder()

	middleware(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {})).ServeHTTP(rr, req)

	assert.Equal(t, http.StatusOK, rr.Code)
}

func TestMiddleware_WithInvalidToken(t *testing.T) {
	secret := []byte("secret")
	middleware := Middleware(secret)

	req, _ := http.NewRequest("GET", "/", nil)
	req.Header.Add("Authorization", "Bearer "+createToken([]byte("wrong secret")))
	rr := httptest.NewRecorder()

	middleware(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {})).ServeHTTP(rr, req)

	assert.Equal(t, http.StatusUnauthorized, rr.Code)
}

func TestMiddleware_WithNoToken(t *testing.T) {
	secret := []byte("secret")
	middleware := Middleware(secret)

	req, _ := http.NewRequest("GET", "/", nil)
	rr := httptest.NewRecorder()

	middleware(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {})).ServeHTTP(rr, req)

	assert.Equal(t, http.StatusOK, rr.Code)
}

func createToken(secret []byte) string {
	token := jwt.New(jwt.SigningMethodHS256)
	ss, _ := token.SignedString(secret)
	return ss
}
